Remove unnecessary mod
